finish the auth for tasks

+++ src/router/tasks.js
@@ -1,9 +1,13 @@
 const express = require('express')
 const Task = require('../models/task')
 const router = express.Router()
+const auth = require('../middleware/auth')
 
-router.post('/tasks', async (req, res) => {
-    const task = new Task(req.body)
+router.post('/tasks', auth, async (req, res) => {
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
     try {
         await task.save()
         res.status(201).send(task)
@@ -13,22 +17,19 @@ router.post('/tasks', async (req, res) => {
     }
 })
 
-router.get('/tasks', async (req, res) => {
+router.get('/tasks', auth, async (req, res) => {
     try {
-        const task = await Task.find({})
-        if(!task) {
-            return res.status(400).send()
-        }
-        res.send(task)
+        await req.user.populate('tasks').execPopulate()
+        res.send(req.user.tasks)
     }
     catch(e){
         res.status(500).send(e)
     }
 })
 
-router.get('/tasks/:id', async (req, res) => {
+router.get('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findOne({_id: req.params.id})
+        const task = await Task.findOne({_id: req.params.id, owner: req.user._id})
         if (!task){
             return res.status(400).send()
         }
@@ -39,7 +40,7 @@ router.get('/tasks/:id', async (req, res) => {
     }
 })
 
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const allowedUpdate = ['description', 'completed']
     const tobeUpdated = Object.keys(req.body)
     const isvalid = tobeUpdated.every(prop => allowedUpdate.includes(prop))
@@ -49,24 +50,24 @@ router.patch('/tasks/:id', async (req, res) => {
     }
 
     try {
-        const task = await Task.findOne({_id: req.params.id})
-        tobeUpdated.forEach(prop => task[prop] = req.body[prop])
-        const res_task = await task.save()
+        const task = await Task.findOne({_id: req.params.id, owner: req.user._id })
     
         if (!task) {
-            return res.status(400).send({error: "Update failed"})
+            return res.status(400).send({})
         }
 
-        res.send(res_task)
+        tobeUpdated.forEach(prop => task[prop] = req.body[prop])
+        await task.save()
+        res.send(task)
     }
     catch(e) {
         res.status(500).send(e)
     }
 })
 
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findOneAndDelete({_id: req.params.id})
+        const task = await Task.findOneAndDelete({_id: req.params.id, owner: req.user._id})
         if (!task){
             return res.status(404).send()
         }
